dash_app/requirements.txt
haversine==2.3.0
dash_html_components==1.1.1
dash_core_components==1.14.1
dash==1.18.1
pandas==1.1.4
Werkzeug==1.0.1
dash_extensions==0.0.42
Flask==1.1.2
dash_bootstrap_components==0.12.0
python_dateutil==2.8.1
